fix deprecated inner constructors in graphmatrices.jl
The inner constructor form:
```julia
	function NormalizedAdjacency(adjmat::CombinatorialAdjacency)
		sf = adjmat.D.^(-1/2)
		return new(adjmat, sf)
	end
 end
```
is replaced by

```julia
 function NormalizedAdjacency{T}(adjmat::CombinatorialAdjacency{T})
  sf = adjmat.D.^(-1/2)
  return NormalizedAdjacency(adjmat, sf)
 end
```

This deprecates the calling convention:

    NormalizedAdjacency{Float64}(adjmat)

in favor of

   NormalizedAdjacency(adjmat)

which is much cleaner. The {Float64} parameter was needed because
the NormalizedAdjacency(CombinatorialAdjacency{Int}(a)) needed to be
of type Float64. This promotion choice was exposed to the caller instead
of being made automatically. Now if you want to construct a
NormalizedAdjacency{Float64} from a CombinatorialAdjacency{Int} you
will have to use the default constructor and do the conversion by hand.